ref(replay): Extract components from `<BreadcrumbItem>`
