+++ static/app/views/replays/detail/breadcrumbs/breadcrumbRow.tsx
@@ -23,12 +23,14 @@ interface Props {
   startTimestampMs: number;
   style: CSSProperties;
   breadcrumbIndex?: number[][];
+  className?: string;
   expandPaths?: string[];
   ref?: React.Ref<HTMLDivElement>;
   updateDimensions?: () => void;
 }
 
 function BreadcrumbRow({
+  className,
   expandPaths,
   frame,
   index,
@@ -63,7 +65,7 @@ function BreadcrumbRow({
   return (
     <BreadcrumbItem
       ref={ref}
-      className={classNames({
+      className={classNames(className, {
         beforeCurrentTime: hasOccurred,
         afterCurrentTime: !hasOccurred,
         beforeHoverTime: currentHoverTime === undefined ? undefined : isBeforeHover,
